+++ lib/wavefront-cli/commands/write.rb
@@ -11,9 +11,9 @@ class WavefrontCommandWrite < WavefrontCommandBase
     ['point [-DnViq] [-c file] [-P profile] [-E proxy] [-t time] ' \
      '[-p port] [-H host] [-T tag...] <metric> <value>',
      'file [-DnViq] [-c file] [-P profile] [-E proxy] [-H host] ' \
-     '[-p port] [-F format] [-m metric] [-T tag...] ' \
-     '[-r rate] <file>']
+     '[-p port] [-F format] [-m metric] [-T tag...] <file>']
   end
+
   def _options
     ['-E, --proxy=URI            proxy endpoint',
      '-t, --time=TIME            time of data point (omit to use ' \
@@ -27,15 +27,13 @@ class WavefrontCommandWrite < WavefrontCommandBase
      'the two will be dot-concatenated, with this value first',
      '-i, --delta                increment metric by given value',
      "-q, --quiet                don't report the points sent summary " \
-     '(unless there were errors)',
-     '-r, --rate=INTEGER         throttle point sending to this many ' \
-     'points per second']
+     '(unless there were errors)']
   end
 
   def postscript
     'Files are whitespace separated, and fields can be defined ' \
-    "with the '-F' option.  Use 't' for timestamp; 'm' for metric " \
-    "name; 'v' for value, 's' for source, and 'T' for tags. Put 'T' " \
+    "with the '-F' option.  Use 't' for timestamp, 'm' for metric " \
+    "name, 'v' for value, 's' for source, and 'T' for tags. Put 'T' " \
     'last.'.cmd_fold(TW, 0)
   end
 end