+++ lib/wavefront-cli/report.rb
@@ -3,7 +3,7 @@ require_relative 'base_write'
 module WavefrontCli
   class Report < BaseWrite
     def send_point(p)
-      wf.write(p)
+      call_write(p)
     rescue Wavefront::Exception::InvalidEndpoint
       abort 'could not speak to API'
     end